Initial structure for a proper FDO compliant icon theme.
The idea here is to provide a "real" fdo icon theme along the icons in the elm edje theme.

To minimize file redundancy the icons are profided directly in a fdo structure and picked
in icon_fdo.edc. Also I'm using all the icons in svg format so that we don't need to provide
all the creazy sizes as per fdo spec.

More icons will comes in the next commits