Revert "tiff: disable by default as it is full of CVE with apparently no chance to get them fixed."
This reverts commit b8860c88f52c7ea3576f88f9399b777646975bd5.

i wouldn't call this full of CVE's:

http://www.cvedetails.com/product/3881/Libtiff-Libtiff.html?vendor_id=2224

i do notice various CVE's on libtiff's mailing list have had patches
committed. the CVE db doesn't track if the CVE has been fixed by
upstream (in an easy to find way) and in which version or on what date so
the CVE db simply is all CVE's since the dawn of time that were ever filed.