Revert "ecore timer - fuix up segv storm that has crept in with frees"
This reverts commit a13570c17c97bb9407c24bcf78ab28eaa4541ad9.

This doesn't really fix the problem which is hidden by eo capability to not
crash on bad unref. With legacy API you are allowed to do a ecore_timer_del
and also return EINA_FALSE. In that case you have a double eo_del (which is
luckily protected) and a double free (that is not). It does crash on the
double free, but the issue is a lifecycle issue. Will bring a better patch
for this.