Elementary toolbar: Don't call "clicked" callback when item is just selected
Summary:
"clicked" callback must be called when an item is really clicked.
But, toolbar had called it whenever an item was selected.
Even, it could be triggered by "elm_toolbar_item_selected_set()" API.
So, "clicked" callback will be called when only an item is clicked.
Also, it changed confusing names of related internal functions.
@fix

Test Plan: N/A

Reviewers: jaehwan, woohyun, cedric

Subscribers: jpeg

Differential Revision: https://phab.enlightenment.org/D3937