autotools: improve libunwind detection
Libuwind may not be shipped with a pkg-config file.
It can be distributed on the system, but the autotools
would fail to detect it because it relied only on pkg-config.

We now first check with pkg-config, and then try to compile and
link a program using libuwind to see if it is supported anyway.

This is a first step towards a working eina_log_backtrace on
Mac OS X.