eina: fix compilation on OpenBSD/Windows of eina_btlog
We only care whether the process terminated successfully or not.
Non-zero values will be anything but a success, so we don't even care
about the exit code of the program.

This should fix compiling on OpenBSD (where WEXITSTATUS needed
to be defined by including a system header) as well on Windows.

Fixes T3733