elm fileselector - make ok/cancel in selector configurable per os
so the theme build can order ok/cancel based on preference for an os,
so make configure have an option for this and build the theme
specifically based on that option. enable the option if you want mac
style cancel, ok or default ok, cancel as is common elsewhere.

  --enable-cancel-ok

is the option

@feature