elm: Partially revert "elementary: avoid random memory access."
This reverts commit ac5a0b7e79d6b32285bd7b794f050826f38cc7c2.

I added assert(pd == sd) in the modified code and couldn't find
anything suspicious. The proper class is used in eo_data_scope_get().
I also can't find a crash or a bug currently in hover. I know there
was a problem earlier, but that was fixed in 3092e08a83f754ba8d49c.

Now, if there is still a problem, we should check it carefully.
Pinging @cedric