FDO icons: and finally install the icons in fdo format
After this commit you will find a new icon theme in your system,
it is called Enlightenment-X and is build using the same icons used
in the elm theme. Using this theme you can have non-efl
and efl apps to use the same icons.

#StupidAutotools that require me to list each icon and each folder
to be able to install them