elementary/colorselector: use layout instead of edje object for components.
Summary:
Issue: As edje object is used instead of layout for components mode,  the content widgets like buttons are added as
subobject of colorselector hence when we unset the edje object on mode change the buttons still remain subobject of colorselector causing issues when all subobjects of colorselector is queried(sub objects are queried during atspi mode).
Soln: Use layout instead of edje object.

Signed-off-by: Shilpa Singh <shilpa.singh@samsung.com>

@fix

Test Plan: Change the mode to Palette mode and access the children of colorselector.

Reviewers: Hermet, cedric

Subscribers: govi, rajeshps, kimcinoo, cedric, jpeg

Differential Revision: https://phab.enlightenment.org/D3987

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>