win: Add property accel_preference
It supports the same syntax as the config value, but
overrides it. This provides a more convenient way to select
the acceleration for a single window.

Its get() value should be what was effective when the window was
created. So it can depend on the config or a previous call
to set().

@feature