evas: (Try to) fix shaders gen without breaking everything
So... adding the .x files to BUILT_SOURCES means they would
get deleted during clean. This was bad, but turned to be awful
when this made the check for writability fail.

See:
8a273158788d630e56b4dc14cdd44f0eaae6b7b5
2b8912462e0ec81efb8e30a63383b811f52e16f5

This commit reverts:
0b69356f1bbceab3261835bc2d4632f8f4669206

Notes:
Please revert if this breaks the build for anyone.
I'm pretty sure the use of top_srcdir is not recommended the
way I've done it.

I hate autofoo.