evas canvas: fix a proxy+map rendering issue.
This fix is wrt the non-updation of the proxy object.
If a source of a proxy object has mapped objects among its children,
Rendering frame won't be correct while they were keeping changing.

The reason is, evas_object_smart_changed_get() couldn't figure those changes
out correctly because of cached-clip visibility.

Cached-clip of smart member can't be used for map surface updation,
because cached-clip includes the visiblity status out of the smart object.

So, loose the criteria to make it rendering smart members in map.