test: Fix strange black rectangle in a test case
elementary_test -to "Entry on Page Scroll"

Click on "Click this and close popup"

A weird rectangle would be temporarily visible while the
popup is fading in. The rectangle was the block_events
rectangle. By default, elm_popup blocks events, and its
parent must be the whole blocked input area. This fixes the
parent in the example.

I think the block_events region should be fading in and out
as well, because it looks terrible right now.

Note: this changes the position of the popup, but makes
the whole "block_events" thing actually work (avoiding
other render artifacts).